services/serving/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
mlflow==2.14.3
pandas==2.2.2
prometheus-client==0.20.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
